test(auth): place recovery diagnostics at page boundary
