+++ turbo/apps/platform/src/views/auth-v2/sign-in/__tests__/sign-in-diagnostics.test.tsx
@@ -1,0 +1,183 @@
+import { fireEvent, screen, waitFor } from "@testing-library/react";
+import { afterAll, beforeEach, describe, expect, it, vi } from "vitest";
+
+import { detachedSetupPage } from "../../../../__tests__/page-helper.ts";
+import {
+  mockAuthV2Capabilities,
+  mockedClerk,
+  mockSignInResource,
+} from "../../../../__tests__/mock-auth.ts";
+import { AUTH_V2_DIAGNOSTIC_EVENT } from "../../../../lib/posthog.ts";
+import { testContext } from "../../../../signals/__tests__/test-helpers.ts";
+
+type Capture = (
+  eventName: string,
+  properties?: Record<string, unknown>,
+) => void;
+
+const { apiOriginMarker, posthog } = vi.hoisted(() => {
+  vi.stubEnv("VITE_POSTHOG_KEY", "phc_auth_v2_recovery_diagnostics_test");
+  window.location.href = "https://app.vm0.ai/";
+  const apiOriginMarker = document.createElement("meta");
+  apiOriginMarker.name = "vm0-api-origin";
+  apiOriginMarker.content = "https://api.vm0.ai";
+  document.head.append(apiOriginMarker);
+  return {
+    apiOriginMarker,
+    posthog: {
+      capture: vi.fn<Capture>(),
+      identify: vi.fn(),
+      init: vi.fn(),
+      register: vi.fn(),
+      reset: vi.fn(),
+      unregister: vi.fn(),
+    },
+  };
+});
+
+vi.mock("posthog-js", () => {
+  return { posthog };
+});
+
+const context = testContext();
+
+afterAll(() => {
+  apiOriginMarker.remove();
+});
+
+beforeEach(() => {
+  posthog.capture.mockClear();
+  posthog.identify.mockClear();
+  posthog.init.mockClear();
+  posthog.register.mockClear();
+  posthog.reset.mockClear();
+  posthog.unregister.mockClear();
+});
+
+function diagnosticCalls(): unknown[][] {
+  return posthog.capture.mock.calls.filter(([eventName]) => {
+    return eventName === AUTH_V2_DIAGNOSTIC_EVENT;
+  });
+}
+
+describe("auth v2 password recovery diagnostics", () => {
+  it.each([
+    {
+      action: "Reset your password",
+      method: "password-reset",
+      operation: "prepare",
+    },
+    {
+      action: "Continue with Apple",
+      method: "apple-oauth",
+      operation: "oauth",
+    },
+    {
+      action: "Email code to p***@example.com",
+      method: "email-code",
+      operation: "prepare",
+    },
+    {
+      action: "Sign in with your passkey",
+      method: "passkey",
+      operation: "passkey",
+    },
+  ] as const)(
+    "attributes the $method recovery action through the rendered page",
+    async ({ action, method, operation }) => {
+      const privateIdentifier = "private.recovery@example.com";
+      const privateProviderCode = `private_${method}_provider_code`;
+      const privateProviderMessage = `Private ${method} provider detail`;
+      const providerFailure = {
+        errors: [
+          {
+            code: privateProviderCode,
+            longMessage: privateProviderMessage,
+          },
+        ],
+      };
+      mockAuthV2Capabilities({ appleOAuth: true, passkey: true });
+      mockSignInResource({
+        identifier: privateIdentifier,
+        status: "needs_first_factor",
+        supportedFirstFactors: [
+          { strategy: "password" },
+          {
+            emailAddressId: "email_private",
+            safeIdentifier: "p***@example.com",
+            strategy: "reset_password_email_code",
+          },
+          { strategy: "oauth_apple" },
+          {
+            emailAddressId: "email_private",
+            safeIdentifier: "p***@example.com",
+            strategy: "email_code",
+          },
+          { strategy: "passkey" },
+        ],
+      });
+      if (operation === "prepare") {
+        mockedClerk.signInPrepareFirstFactor.mockRejectedValueOnce(
+          providerFailure,
+        );
+      } else if (operation === "oauth") {
+        mockedClerk.signInAuthenticateWithRedirect.mockRejectedValueOnce(
+          providerFailure,
+        );
+      } else {
+        mockedClerk.signInAuthenticateWithPasskey.mockRejectedValueOnce(
+          providerFailure,
+        );
+      }
+
+      const path = "/v2/sign-in";
+      context.mocks.browser.url(`https://app.vm0.ai${path}`);
+      detachedSetupPage({
+        context,
+        path,
+        session: null,
+        user: null,
+      });
+
+      fireEvent.click(
+        await screen.findByRole("button", { name: "Forgot password?" }),
+      );
+      await expect(
+        screen.findByRole("heading", { name: "Forgot Password?" }),
+      ).resolves.toBeVisible();
+      fireEvent.click(await screen.findByRole("button", { name: action }));
+
+      const alert = await screen.findByRole("alert");
+      expect(alert).toHaveTextContent(
+        "This action couldn't be completed. Please try again later or contact support if this persists.",
+      );
+      expect(alert).not.toHaveTextContent(privateProviderMessage);
+      await waitFor(() => {
+        expect(document.activeElement).toBe(alert);
+      });
+      await waitFor(() => {
+        expect(diagnosticCalls()).toStrictEqual([
+          [
+            AUTH_V2_DIAGNOSTIC_EVENT,
+            {
+              error_category: "provider-error",
+              flow: "sign-in",
+              method,
+              outcome: "failure",
+              step: "recovery",
+            },
+          ],
+        ]);
+      });
+      const serializedCalls = JSON.stringify(diagnosticCalls());
+      for (const prohibitedValue of [
+        privateIdentifier,
+        privateProviderCode,
+        privateProviderMessage,
+        "email_private",
+      ]) {
+        expect(serializedCalls).not.toContain(prohibitedValue);
+      }
+    },
+  );
+});
